fix admin controller and add view user profile

+++ backend/src/main/java/com/chattingweb/backend/services/admin/DeletePost.java
@@ -5,13 +5,13 @@ import jakarta.validation.constraints.Size;
 
 public class DeletePost {
     @NotBlank(message = "post ID is required")
-    private Integer ID;
+    private Long ID;
 
     @NotBlank(message = "Delete reason is required")
     @Size(max = 200,message = "Length of ban reason must be below 200")
     private String deleteReason;
 
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
     public String getdeleteReason() {
